+++ FS.FilterExpressionCreator.Tests/Tests/TypeFilter/FilterForGuidBySyntaxTests.cs
@@ -19,7 +19,7 @@ namespace FS.FilterExpressionCreator.Tests.Tests.TypeFilter
         private static readonly TestModel<Guid>[] _testItems = {
             new() { ValueA = Guid.Parse("df72ce74-686c-4c0f-a11f-5c8e50a213ab") },
             new() { ValueA = Guid.Parse("6cda682c-e7ff-43e8-b4d9-f8b27a7d62f2") },
-            new() { ValueA = Guid.Parse("9e3aada6-c0cc-4d4d-bea3-7058777e98d1") },
+            new() { ValueA = Guid.Parse("6cda682c-c0cc-4d4d-bea3-7058777e98d1") },
         };
 
         // ReSharper disable RedundantExplicitArrayCreation
@@ -29,7 +29,8 @@ namespace FS.FilterExpressionCreator.Tests.Tests.TypeFilter
             FilterTestCase.Create<Guid>(1102, "6cda682c-e7ff-43e8-b4d9-f8b27a7d62f2", x => x == Guid.Parse("6cda682c-e7ff-43e8-b4d9-f8b27a7d62f2")),
             FilterTestCase.Create<Guid>(1103, "6CDA682C-E7FF-43E8-B4D9-F8B27A7D62F2", x => x == Guid.Parse("6cda682c-e7ff-43e8-b4d9-f8b27a7d62f2")),
 
-            FilterTestCase.Create<Guid>(1200, "~null", new FilterExpressionCreationException("Filter operator 'Contains' not allowed for property type 'System.Guid'")),
+            FilterTestCase.Create<Guid>(1200, "~686c", x => x == Guid.Parse("df72ce74-686c-4c0f-a11f-5c8e50a213ab")),
+            FilterTestCase.Create<Guid>(1201, "~6cda682c", x => x == Guid.Parse("6cda682c-e7ff-43e8-b4d9-f8b27a7d62f2") || x == Guid.Parse("6cda682c-c0cc-4d4d-bea3-7058777e98d1")),
 
             FilterTestCase.Create<Guid>(1300, "=null", new FilterExpressionCreationException("Unable to parse given filter value")),
             FilterTestCase.Create<Guid>(1301, "=", new FilterExpressionCreationException("Unable to parse given filter value")),